Movie/movie_booking_ai_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.7
python-decouple==3.8
razorpay==1.4.2
stripe==7.8.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
openai==1.3.7
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0
freezegun==1.2.2
